Modify _copy to unlink target file before copying
Follow the behaviour of utilities like install(1) to unlink the target
file before copying. If a process has a handle on the target file i.e. a
shared library that's been mapped into memory, overwriting the contents
of the file may result in invalid memory memory references as regions of
the mapped file are read in on demand.

Unlinking keeps the contents of the file around as long as it's
opened, but allows us to create a new filesystem reference to the target
file.